# plugins/orchestrator-common/src/openapi_types.ts
import { components } from '../api/models/schema';

export type WorkflowOverviewListResultDTO =
  components['schemas']['WorkflowOverviewListResultDTO'];
export type WorkflowOverviewDTO = components['schemas']['WorkflowOverviewDTO'];
export type WorkflowDTO = components['schemas']['WorkflowDTO'];
export type WorkflowListResultDTO =
  components['schemas']['WorkflowListResultDTO'];
export type ProcessInstanceDTO = components['schemas']['ProcessInstanceDTO'];
export type ProcessIntancesDTO = components['schemas']['ProcessInstancesDTO'];
export type ProcessInstanceStatusDTO2 =
  components['schemas']['ProcessInstanceStatusDTO'];
export type WorkflowRunStatusDTO =
  components['schemas']['WorkflowRunStatusDTO'];